[AIP-21] Move contrib.operators.s3_to_gcs_operator providers.google.cloud.operators.s3_to_gcs
