+++ scripts/ci/pylint_todo.txt
@@ -52,7 +52,7 @@
 ./airflow/providers/amazon/aws/operators/s3_copy_object.py
 ./airflow/providers/amazon/aws/operators/s3_delete_objects.py
 ./airflow/providers/amazon/aws/operators/s3_list.py
-./airflow/contrib/operators/s3_to_gcs_operator.py
+./airflow/providers/google/cloud/operators/s3_to_gcs.py
 ./airflow/providers/amazon/aws/operators/s3_to_sftp.py
 ./airflow/providers/amazon/aws/operators/sagemaker_base.py
 ./airflow/providers/amazon/aws/operators/sagemaker_endpoint_config.py
